requirements.txt
twilio
gunicorn
flask
requests
python-dotenv
vosk
soundfile
cryptography
google-generativeai
